Move DotDict helper class to a more central location for use elsewhere
