+++ SpiffWorkflow/dmn/engine/DMNEngine.py
@@ -1,18 +1,11 @@
 import logging
 from argparse import Namespace
 from collections import namedtuple
-
+from ...operators import DotDict
 from decimal import Decimal
 import datetime
 
 
-class DotDict(dict):
-    """dot.notation access to dictionary attributes"""
-    def __getattr__(*args):
-        val = dict.get(*args)
-        return DotDict(val) if type(val) is dict else val
-    __setattr__ = dict.__setitem__
-    __delattr__ = dict.__delitem__
 
 class DMNEngine:
     """
