+++ SpiffWorkflow/operators.py
@@ -31,6 +31,15 @@ class Term(object):
     """
     pass
 
+class DotDict(dict):
+    """dot.notation access to dictionary attributes"""
+    def __getattr__(*args):
+        val = dict.get(*args)
+        return DotDict(val) if type(val) is dict else val
+    __setattr__ = dict.__setitem__
+    __delattr__ = dict.__delitem__
+
+
 
 class Attrib(Term):
